refactor: move AddAttachmentModalCarouselView to separate component
