+++ src/pages/media/AttachmentModalScreen/routes/report/ReportAddAttachmentModalContent/AddAttachmentModalCarouselView.tsx
@@ -1,0 +1,62 @@
+import React, {useEffect, useMemo, useState} from 'react';
+import AttachmentCarouselView from '@components/Attachments/AttachmentCarousel/AttachmentCarouselView';
+import useCarouselArrows from '@components/Attachments/AttachmentCarousel/useCarouselArrows';
+import useAttachmentErrors from '@components/Attachments/AttachmentView/useAttachmentErrors';
+import type {Attachment} from '@components/Attachments/types';
+import type {AttachmentContentProps} from '@pages/media/AttachmentModalScreen/AttachmentModalBaseContent/types';
+import type {FileObject} from '@pages/media/AttachmentModalScreen/types';
+
+const convertFileToAttachment = (file: FileObject | undefined): Attachment => {
+    if (!file) {
+        return {source: ''};
+    }
+
+    return {
+        file,
+        source: file.uri ?? '',
+    };
+};
+
+function AddAttachmentModalCarouselView({fileToDisplay, files}: AttachmentContentProps) {
+    const {setAttachmentError, clearAttachmentErrors} = useAttachmentErrors();
+    const {shouldShowArrows, setShouldShowArrows, autoHideArrows, cancelAutoHideArrows} = useCarouselArrows();
+
+    const [page, setPage] = useState<number>(0);
+    const attachments = useMemo(() => {
+        if (Array.isArray(files)) {
+            return files?.map((file) => convertFileToAttachment(file)) ?? [];
+        }
+
+        if (!files) {
+            return [];
+        }
+
+        return [convertFileToAttachment(files)];
+    }, [files]);
+    const currentAttachment = useMemo(() => convertFileToAttachment(fileToDisplay), [fileToDisplay]);
+
+    useEffect(() => {
+        clearAttachmentErrors();
+    }, [clearAttachmentErrors]);
+
+    if (attachments.length === 0 || !currentAttachment) {
+        return null;
+    }
+
+    return (
+        <AttachmentCarouselView
+            attachments={attachments}
+            source={currentAttachment.source}
+            page={page}
+            setPage={setPage}
+            autoHideArrows={autoHideArrows}
+            cancelAutoHideArrow={cancelAutoHideArrows}
+            setShouldShowArrows={setShouldShowArrows}
+            onAttachmentError={setAttachmentError}
+            shouldShowArrows={shouldShowArrows}
+        />
+    );
+}
+AddAttachmentModalCarouselView.displayName = 'AddAttachmentModalCarouselView';
+
+export default AddAttachmentModalCarouselView;
